rename appointments to company_appointments
