+++ datapackage_pipelines_budgetkey/pipelines/people/company_appointments/maya/scrape-individual-nomination.py
@@ -3,7 +3,7 @@ import requests
 
 from datapackage_pipelines.wrapper import process
 from datapackage_pipelines_budgetkey.common.object_storage import object_storage
-from datapackage_pipelines_budgetkey.pipelines.people.appointments.maya.maya_nomination_form import MayaForm, ParseError
+from datapackage_pipelines_budgetkey.pipelines.people.company_appointments.maya.maya_nomination_form import MayaForm, ParseError
 
 logging.getLogger("requests").setLevel(logging.WARNING)
 logging.getLogger("urllib3").setLevel(logging.WARNING)
