+++ datapackage_pipelines_budgetkey/pipelines/people/company_appointments/media/themarker/pipeline-spec.yaml
@@ -57,6 +57,4 @@ themarker:
     -
       run: dump.to_path
       parameters:
-          out-path: /var/datapackages/people/appointments/media/themarker
-
-
+          out-path: /var/datapackages/people/company_appointments/media/themarker
