+++ datapackage_pipelines_budgetkey/pipelines/people/company_appointments/pipeline-spec.yaml
@@ -1,8 +1,8 @@
 all-appointments:
   dependencies:
-    - pipeline: ./people/appointments/media/themarker/themarker
-    - pipeline: ./people/appointments/media/calcalist/calcalist
-    - pipeline: ./people/appointments/maya/maya-nominations
+    - pipeline: ./people/company_appointments/media/themarker/themarker
+    - pipeline: ./people/company_appointments/media/calcalist/calcalist
+    - pipeline: ./people/company_appointments/maya/maya-nominations
 
   pipeline:
     -
@@ -14,19 +14,19 @@ all-appointments:
     -
       run: load_resource
       parameters:
-        url: /var/datapackages/people/appointments/media/themarker/datapackage.json
+        url: /var/datapackages/people/company_appointments/media/themarker/datapackage.json
         resource: 0
 
     -
       run: load_resource
       parameters:
-        url: /var/datapackages/people/appointments/media/calcalist/datapackage.json
+        url: /var/datapackages/people/company_appointments/media/calcalist/datapackage.json
         resource: nominations-list
 
     -
       run: load_resource
       parameters:
-        url: /var/datapackages/people/appointments/maya/maya-nominations/datapackage.json
+        url: /var/datapackages/people/company_appointments/maya/maya-nominations/datapackage.json
         resource: maya-nominations
 
     -
@@ -127,4 +127,4 @@ all-appointments:
     -
       run: dump.to_path
       parameters:
-          out-path: /var/datapackages/people/appointments/all
+          out-path: /var/datapackages/people/company_appointments/all
